feat: auth, dataPool, course media, purchase & subscription services
- Add video preview timestamps migration
- Add course payment model migration
- Update auth service and controller with improved session handling
- Expand dataPool service and controller for course enrollment and video streaming
- Update purchase and subscription services with Razorpay webhook handling
- Update S3 service for course media management
- Update email service with SES cross-account role support
- Improve validators, rate limiter, and env config

+++ src/utils/validators.ts
@@ -82,16 +82,19 @@ export const createPlanSchema = z.object({
   discountLabel: z.string().max(100).optional(),
 
   // Pricing Overrides
+  // Note: *Value fields use .number() not .number().int() because PERCENT mode accepts
+  // decimals (e.g. GST at 18.5%). FLAT mode values are paise integers but Zod validation
+  // here is intentionally permissive — the service layer enforces rounding for FLAT mode.
   gstMode: z.enum(['PERCENT', 'FLAT']).nullable().optional(),
-  gstValue: z.number().int().min(0).nullable().optional(),
+  gstValue: z.number().min(0).nullable().optional(),
   gatewayChargeMode: z.enum(['PERCENT', 'FLAT']).nullable().optional(),
-  gatewayChargeValue: z.number().int().min(0).nullable().optional(),
+  gatewayChargeValue: z.number().min(0).nullable().optional(),
   serviceFeeMode: z.enum(['PERCENT', 'FLAT']).nullable().optional(),
-  serviceFeeValue: z.number().int().min(0).nullable().optional(),
+  serviceFeeValue: z.number().min(0).nullable().optional(),
   processingFeeMode: z.enum(['PERCENT', 'FLAT']).nullable().optional(),
-  processingFeeValue: z.number().int().min(0).nullable().optional(),
+  processingFeeValue: z.number().min(0).nullable().optional(),
   platformChargeMode: z.enum(['PERCENT', 'FLAT']).nullable().optional(),
-  platformChargeValue: z.number().int().min(0).nullable().optional(),
+  platformChargeValue: z.number().min(0).nullable().optional(),
 
   // Pricing for each billing cycle
   pricing: z.array(planPricingSchema).default([]),
@@ -153,17 +156,17 @@ export const updatePlanSchema = z.object({
   discountPercent: z.number().min(0).max(100).optional().nullable(),
   discountLabel: z.string().max(100).optional().nullable(),
 
-  // Pricing Overrides
+  // Pricing Overrides (see createPlanSchema for why .int() is omitted)
   gstMode: z.enum(['PERCENT', 'FLAT']).nullable().optional(),
-  gstValue: z.number().int().min(0).nullable().optional(),
+  gstValue: z.number().min(0).nullable().optional(),
   gatewayChargeMode: z.enum(['PERCENT', 'FLAT']).nullable().optional(),
-  gatewayChargeValue: z.number().int().min(0).nullable().optional(),
+  gatewayChargeValue: z.number().min(0).nullable().optional(),
   serviceFeeMode: z.enum(['PERCENT', 'FLAT']).nullable().optional(),
-  serviceFeeValue: z.number().int().min(0).nullable().optional(),
+  serviceFeeValue: z.number().min(0).nullable().optional(),
   processingFeeMode: z.enum(['PERCENT', 'FLAT']).nullable().optional(),
-  processingFeeValue: z.number().int().min(0).nullable().optional(),
+  processingFeeValue: z.number().min(0).nullable().optional(),
   platformChargeMode: z.enum(['PERCENT', 'FLAT']).nullable().optional(),
-  platformChargeValue: z.number().int().min(0).nullable().optional(),
+  platformChargeValue: z.number().min(0).nullable().optional(),
 
   // Pricing (replaces all existing pricing when provided)
   pricing: z.array(planPricingSchema).optional(),
